requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx>=0.24,<0.26
supabase==2.3.4
pydantic==2.6.1
python-multipart==0.0.9
resend==2.19.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pytest==8.3.5
pytest-asyncio==0.24.0